fix(#83) :: 사이드바 컴포넌트화
