+++ src/app/(admin)/admin/page.tsx
@@ -1,90 +1,13 @@
 'use client';
 
-import React, { useState } from 'react';
-import { useRouter } from 'next/navigation';
+import Sidebar from '@/components/Admin/Sidebar';
+import React from 'react';
 
 const AdminHome = () => {
-  const router = useRouter();
-  const [activeMenu, setActiveMenu] = useState('dashboard');
-
-  const menuItems = [
-    { id: 'dashboard', label: '대시보드', icon: '📊' },
-    { id: 'users', label: '유저 관리', icon: '👥' },
-    { id: 'inventory', label: '재고 관리', icon: '📦' },
-    { id: 'products', label: '상품 관리', icon: '🛍️' },
-    { id: 'orders', label: '주문 관리', icon: '📋' },
-    { id: 'analytics', label: '분석', icon: '📈' },
-  ];
-
-  const handleMenuClick = (menuId: string) => {
-    setActiveMenu(menuId);
-
-    // 라우팅 처리
-    switch (menuId) {
-      case 'users':
-        router.push('/admin/user');
-        break;
-      case 'inventory':
-        router.push('/admin/inventory');
-        break;
-      case 'products':
-        router.push('/admin/products');
-        break;
-      case 'orders':
-        router.push('/admin/orders');
-        break;
-      case 'analytics':
-        router.push('/admin/analytics');
-        break;
-      default:
-        break;
-    }
-  };
-
   return (
     <div className="flex h-screen bg-gray-900">
       {/* 사이드바 */}
-      <div className="flex flex-col w-64 text-white bg-gray-800">
-        {/* 로고 */}
-        <div className="p-6 border-b border-gray-700">
-          <h1 className="text-xl font-bold text-blue-400">Chaellimi</h1>
-          <p className="text-sm text-gray-400">관리자 패널</p>
-        </div>
-
-        {/* 메뉴 */}
-        <nav className="flex-1 px-4 py-6">
-          <ul className="space-y-2">
-            {menuItems.map((item) => (
-              <li key={item.id}>
-                <button
-                  onClick={() => handleMenuClick(item.id)}
-                  className={`w-full flex items-center px-4 py-3 rounded-lg transition-colors ${
-                    activeMenu === item.id
-                      ? 'bg-blue-600 text-white'
-                      : 'text-gray-300 hover:bg-gray-700 hover:text-white'
-                  }`}
-                >
-                  <span className="mr-3 text-lg">{item.icon}</span>
-                  {item.label}
-                </button>
-              </li>
-            ))}
-          </ul>
-        </nav>
-
-        {/* 하단 */}
-        <div className="p-4 border-t border-gray-700">
-          <div className="flex items-center">
-            <div className="flex items-center justify-center w-8 h-8 bg-blue-600 rounded-full">
-              <span className="text-sm font-bold">A</span>
-            </div>
-            <div className="ml-3">
-              <p className="text-sm font-medium">Admin</p>
-              <p className="text-xs text-gray-400">관리자</p>
-            </div>
-          </div>
-        </div>
-      </div>
+      <Sidebar activeMenu="dashboard" />
 
       {/* 메인 컨텐츠 */}
       <div className="flex flex-col flex-1">
@@ -112,7 +35,7 @@ const AdminHome = () => {
           <div className="grid grid-cols-1 gap-6 mb-8 md:grid-cols-2 lg:grid-cols-4">
             {/* 유저 관리 카드 */}
             <div
-              onClick={() => handleMenuClick('users')}
+              onClick={() => (window.location.href = '/admin/user')}
               className="p-6 text-white transition-shadow cursor-pointer bg-gradient-to-br from-blue-500 to-blue-600 rounded-xl hover:shadow-lg"
             >
               <div className="flex items-center justify-between">
@@ -129,7 +52,7 @@ const AdminHome = () => {
 
             {/* 재고 관리 카드 */}
             <div
-              onClick={() => handleMenuClick('inventory')}
+              onClick={() => (window.location.href = '/admin/inventory')}
               className="p-6 text-white transition-shadow cursor-pointer bg-gradient-to-br from-green-500 to-green-600 rounded-xl hover:shadow-lg"
             >
               <div className="flex items-center justify-between">
@@ -146,7 +69,7 @@ const AdminHome = () => {
 
             {/* 상품 관리 카드 */}
             <div
-              onClick={() => handleMenuClick('products')}
+              onClick={() => (window.location.href = '/admin/product')}
               className="p-6 text-white transition-shadow cursor-pointer bg-gradient-to-br from-purple-500 to-purple-600 rounded-xl hover:shadow-lg"
             >
               <div className="flex items-center justify-between">
@@ -160,23 +83,6 @@ const AdminHome = () => {
                 </div>
               </div>
             </div>
-
-            {/* 주문 관리 카드 */}
-            <div
-              onClick={() => handleMenuClick('orders')}
-              className="p-6 text-white transition-shadow cursor-pointer bg-gradient-to-br from-orange-500 to-orange-600 rounded-xl hover:shadow-lg"
-            >
-              <div className="flex items-center justify-between">
-                <div>
-                  <p className="text-sm text-orange-100">주문 관리</p>
-                  <p className="text-2xl font-bold">345</p>
-                  <p className="text-sm text-orange-100">총 주문</p>
-                </div>
-                <div className="flex items-center justify-center w-12 h-12 bg-orange-400 rounded-full">
-                  <span className="text-2xl">📋</span>
-                </div>
-              </div>
-            </div>
           </div>
 
           {/* 빠른 액세스 섹션 */}
@@ -186,7 +92,7 @@ const AdminHome = () => {
             </h3>
             <div className="grid grid-cols-2 gap-4 md:grid-cols-4">
               <button
-                onClick={() => handleMenuClick('users')}
+                onClick={() => (window.location.href = '/admin/user')}
                 className="p-4 transition-colors rounded-lg bg-blue-50 hover:bg-blue-100"
               >
                 <div className="text-center">
@@ -196,7 +102,7 @@ const AdminHome = () => {
               </button>
 
               <button
-                onClick={() => handleMenuClick('inventory')}
+                onClick={() => (window.location.href = '/admin/inventory')}
                 className="p-4 transition-colors rounded-lg bg-green-50 hover:bg-green-100"
               >
                 <div className="text-center">
@@ -206,7 +112,7 @@ const AdminHome = () => {
               </button>
 
               <button
-                onClick={() => handleMenuClick('products')}
+                onClick={() => (window.location.href = '/admin/product')}
                 className="p-4 transition-colors rounded-lg bg-purple-50 hover:bg-purple-100"
               >
                 <div className="text-center">
@@ -216,7 +122,7 @@ const AdminHome = () => {
               </button>
 
               <button
-                onClick={() => handleMenuClick('analytics')}
+                onClick={() => (window.location.href = '/admin/analytics')}
                 className="p-4 transition-colors rounded-lg bg-yellow-50 hover:bg-yellow-100"
               >
                 <div className="text-center">
